Old aggregation function clean up (step 1)
Refactor new aggregation functions to contain all necessary methods for the clean up.
Add generic to new aggregation function interface.
Remove ResultHolderFactory, add getResultHolder api into aggregation function interface.
Remove all huge switch on aggregation function name, let each aggregation function handles them.

+++ pinot-core/src/main/java/com/linkedin/pinot/core/common/DataFetcher.java
@@ -369,7 +369,7 @@ public class DataFetcher {
    * @param outValues value array buffer.
    * @param outStartPos output start position.
    */
-  public void fetchHashCodes(String column, int[] inDocIds, int inStartPos, int length, double[] outValues, int outStartPos) {
+  public void fetchHashCodes(String column, int[] inDocIds, int inStartPos, int length, int[] outValues, int outStartPos) {
     Dictionary dictionary = getDictionaryForColumn(column);
     fetchSingleDictIds(column, inDocIds, inStartPos, length, _reusableDictIds, 0);
 
@@ -388,7 +388,7 @@ public class DataFetcher {
    * @param outValues value array buffer.
    * @param outStartPos output start position.
    */
-  public void fetchHashCodes(String column, int[] inDocIds, int inStartPos, int length, double[][] outValues, int outStartPos) {
+  public void fetchHashCodes(String column, int[] inDocIds, int inStartPos, int length, int[][] outValues, int outStartPos) {
     Dictionary dictionary = getDictionaryForColumn(column);
     BlockMultiValIterator iterator = (BlockMultiValIterator) getBlockValIteratorForColumn(column);
 
@@ -396,8 +396,7 @@ public class DataFetcher {
     for (int i = inStartPos; i < inEndPos; i++, outStartPos++) {
       iterator.skipTo(inDocIds[i]);
       int numValues = iterator.nextIntVal(_reusableMVDictIds);
-
-      outValues[outStartPos] = new double[numValues];
+      outValues[outStartPos] = new int[numValues];
       for (int j = 0; j < numValues; j++) {
         outValues[outStartPos][j] = dictionary.get(_reusableMVDictIds[j]).hashCode();
       }
